Consensus: MOVEONLY: Move a lot of consensus functions from main.cpp to consensus.cpp
