+++ src/test/checkblock_tests.cpp
@@ -6,7 +6,6 @@
 #include "clientversion.h"
 #include "consensus/consensus.h"
 #include "consensus/validation.h"
-#include "main.h" // For CheckBlock
 #include "primitives/block.h"
 #include "test/test_bitcoin.h"
 #include "utiltime.h"
